refactor: settings components split
